upsc_essay_app/requirements.txt
streamlit>=1.28.0
langgraph>=0.0.55
langchain-openai>=0.1.9
langchain-huggingface>=0.0.3
python-dotenv>=1.0.0
pydantic>=2.7.4
typing-extensions>=4.8.0
